key_service/requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.1

# Cryptography
pycryptodome==3.20.0
cryptography==43.0.1

# Database drivers
psycopg2-binary==2.9.9   # PostgreSQL
redis==5.0.3             # Redis client

# Monitoring & metrics
prometheus-client==0.20.0
psutil==6.0.0

# Utilities
python-jose==3.3.0       # JWT signing/verification
passlib[bcrypt]==1.7.4   # Password hashing
